Use a bit array to make groups behave more like dwm tags
In this commit, sticky_windows and all groups are removed.

To replace the group functions, three functions are in place.

* group_switch will map clients whose group is n, unmap otherwise.
* group_combine will map all clients whose group is n.
* group_combine_or_toggle will map all clients whose group is n or unmap if group n is already active.

Instead of sticky_windows, each new client's group will be set to the "main group" configuration. This works akin to dwm's tag system!

Man page is yet to be updated.